sbomnix: Add --depth command-line option
Add --depth option to sbomnix to set the depth of the included
dependencies.

As an example, --depth=1 indicates the SBOM should include only
the NIX_PATH direct dependencies. With --depth=2, the output SBOM
includes the direct dependencies and the first level of transitive
dependencies. By default, when --depth is not specified, the output
SBOM includes all dependencies all the way to the root of the
dependency tree.

Signed-off-by: Henri Rosten <henri.rosten@unikie.com>