nix/packages: make reuse a python library
Not using the pip install mechanism exposed another problem: sbomdb.py
actually imports a private module from reuse.

Reuse however is not packaged as an element of the python package set,
but as an application, causing it to be not in PYTHONPATH of the python
composed with python3.withPackages.

Fix this by packaging it as a python package, not an application.
This fixes the pylint unknown import error, or broken execution inside
the devshell for this package.

Signed-off-by: Florian Klink <flokli@flokli.de>