sbomnix: Generate stable UUIDs for SBOMs
Using a random UUID means that the output is not reproducible.

Having reproducible outputs means verifiable output. This can also help
with workflows using the contents of a document to determine if a
particular task needs to be done again.

We already have a unique identifier [at home]. It's the Nix store path.

The CycloneDX JSON reference states:

> Every BOM generated SHOULD have a unique serial number, even if the
> contents of the BOM have not changed over time.

The word SHOULD is interpreted as per RFC2119. The properties of Nix are
in my opinion particular circumstances allowing us to ignore this
particular item. We know none of the inputs have changed, so we are
replaying *an existing SBOM*.

Given that the reproducibility properties of the paths the SBOM refers
to *don't matter in the slightest*, we know for a fact that we are not
generating a "new" SBOM. This is the same SBOM, but built again.

Though, this is ignoring the metadata and CPE information, which can and
do change depending on how the data was gathered by sbomnix. This would
in turn change the output in a way where whilt the *contents of the BOM*
have not changed, the *data presented by the BOM* did.

While this helps with reproducibility, it is not sufficient by itself.
Ignoring the metadata and CPE information, the timestamp field is a
source of reproducibility issues. Using `faketime` can work around that.

A suggested way forward to handle the additional metadata would be to
add some of the BOM data into the *name* value that is encoded in the
document's UUID. In fact, adding *all* of the BOM's data would be
another way to properly handle this problem. For example if the CPE
database is updated, but none of the guessed CPEs in the SBOM change,
the UUID would be left unchanged. When any CPE change from such an
update, the UUID would naturally be changed.

And since the contents of the SBOM *naturally* follows similar
semantics as the store path, encoding all the inputs, it would be
sufficient to only hash the contents + metadata, as long as it's in a
stable order.

Signed-off-by: Samuel Dionne-Riel <samuel.dionne-riel@cyberus-technology.de>