sbomnix: Use stable UUID only when plausible
This removes some of the ambiguity that could be produced by the
previous change, but not all. The CPE list is still an unstable input
that may change the resulting output.

A follow-up change will allow disabling using the CPE matching
heuristics, which also help with sandboxed build: getting the CPE
"database" is not permitted.

Signed-off-by: Samuel Dionne-Riel <samuel.dionne-riel@cyberus-technology.de>