sbomnix: Add `--exclude-cpe-matching` flag
This results in two useful new behaviours:

 - Makes the UUIDv5 serial usage actually stable
 - Skips downloading CPE information

The latter is useful in situations where downloading the database used
for matching CPEs is not possible. For example, when building the SBOM
during a Nix build, in the Nix sandbox.

Note that even if we were to add a `--cpe-database-file <database.csv>`
option, it would still be useful to skip this step entirely: it may
provide different data for a given SBOM when the database changes.

For building an SBOM programatically as part of a Nix build, it would be
*generally undesirable*, assuming perfect reproducibility for a given
input is preferred.

Signed-off-by: Samuel Dionne-Riel <samuel.dionne-riel@cyberus-technology.de>